add 4.21*

+++ part4/blogilista/controllers/blogs.js
@@ -2,7 +2,6 @@ const blogsRouter = require('express').Router();
 const jwt = require('jsonwebtoken');
 const Blog = require('../models/blog');
 const User = require('../models/user');
-const logger = require('../utils/logger');
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -54,9 +53,25 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  logger.info(`id to DELETE: ${request.params.id}`);
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const token = request.token;
+
+  if (!token) {
+    return response.status(401).json({ error: 'token missing' });
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: 'permission denied' });
+  }
+
 });
 
 blogsRouter.put('/:id', async (request, response) => {
